[tsan] Shadow memory setup for OS X
Updating the shadow memory initialization in `tsan_platform_mac.cc` to also initialize the meta shadow and to mprotect the memory ranges that need to be avoided.

Differential Revision: http://reviews.llvm.org/D14324



git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@252044 91177308-0d34-0410-b5e6-96231b3b80d8
